+++ src/partials/take_order.html
@@ -13,7 +13,7 @@
         <div class="form__field">
             <label for="email" class="form__name"></label>
             <input type="email" class="form__input" name="mail" id="mail" placeholder="Email"/>
-                
+             
             </div>
             <div class="form__field">
                 <select name="" class="form__input">
@@ -44,9 +44,14 @@
                </div>
                </div>
                 <div class="form__field">
-                    <label for="message" class="form__name">Leave your message</label>
-                    <textarea name="comment" id="comment" rows="10" placeholder=""></textarea>
+                    <div>
+                        <label for="message" class="form__name">Leave your message</label>
                     </div>
+                    
+                    <div>
+                        <textarea name="comment" id="comment" rows="10" placeholder=""></textarea>
+                    </div>
+                </div>
                
                     <button class="button button-buy" type="submit" data-modal__buy-open>buy it now</button>
