+++ src/sass/components/_modal.scss
@@ -16,6 +16,7 @@
   visibility: visible;
   z-index: 9;
   overflow-y: scroll;
+  background-color: rgba(0, 20, 20, 0.8);
 
   transition: opacity 250ms cubic-bezier(0.4, 0, 0.2, 1);
 }
@@ -24,7 +25,7 @@
   top: 50%;
   left: 50%;
   padding: 40px;
-  width: 100%;
+  width: 272px;
   min-height: 367px;
   border-radius: 4px;
   overflow: auto;
@@ -35,28 +36,47 @@
 
   @media screen and (min-width: $tablet) {
     width: 688px;
+    height: 500px;
     top: 50%;
     left: 50%;
+    margin-left: auto;
+    margin-right: auto;
   }
   @media screen and (min-width: $desktop) {
     width: 768px;
     height: 524px;
     top: 50%;
     left: 50%;
+    margin-left: auto;
+    margin-right: auto;
   }
 }
 .modal__title {
   margin-top: 60px;
-  width: 214px;
+  margin-bottom: 32px;
   font-weight: 500;
   font-size: 24px;
   line-height: 1.17;
   text-align: center;
   letter-spacing: 0.03em;
   color: #ffffff;
+
+  @media screen and (min-width: $tablet) {
+    
+    margin-bottom: 56px;
+    font-size: 56px;
+    line-height: 1.18;
+    text-align: center;
+  }
+  @media screen and (min-width: $desktop) {
+    font-size: 64px;
+    line-height: 1.18;
+    text-align: center;
+    margin-bottom: 56px;
+    
+  }
 }
 .button-shopping {
-  margin-top: 32px;
     transition: color 250ms cubic-bezier(0.4, 0, 0.2, 1);
 }
 .close-btn {
@@ -75,11 +95,18 @@
   background-color: $header-background-color;
 
   transition: color 250ms cubic-bezier(0.4, 0, 0.2, 1);
+  &:hover {
+    box-shadow: 0px 8px 20px rgba(134, 229, 67, 0.5),
+      0px 4px 10px rgba(255, 255, 255, 0.1);
+  }
 }
 
 .close-btn__icon {
   .close-btn:hover &,
   .close-btn:focus & {
-    fill: #ffffff;
+    &:hover {
+      box-shadow: 0px 8px 20px rgba(134, 229, 67, 0.5),
+        0px 4px 10px rgba(255, 255, 255, 0.1);
+    }
   }
 }
